Added enum count entries to `KeyboardKey`, `MouseButton`, and `GamepadButton`
This allows iterating over all possible inputs from the keyboard, mouse, or gamepad more easily without lots of copy pasting. The `MOUSE_BUTTON_NUM_BUTTONS`, `GAMEPAD_BUTTON_NUM_BUTTONS`, and `GAMEPAD_AXIS_NUM_AXIS` entries are robust to the potential future addition of more entries in these enums because they are placed at the end of the enum declaration.  For the `KeyboardKey` enum I used the value of 512. This enum is out of order and so simply placing the `KEY_NUM_KEYS` enum at the end wouldn't be robust to change. Instead I used the value of the `#define MAX_KEYBOARD_KEYS = 512` found in rcore.c used for the length of the internal arrays that store the Keyboard state.