REMOVED: Sponsors doc
This file requires manual maintenance and continuous review, it also changes from day to day moving sponsors from present-to-past and sometimes the other way round.

GitHub already lists the sponsors on the user profile, in a better an more visual way.

I also have doubts this should be a document distributed with raylib sources...