Win32: resolve some symbols re-definition of windows.h in glfw3native.h
This reflects GLFW's fix: https://github.com/glfw/glfw/issues/1348

This enables to build with a external GLFW containing the
following fix:

* https://github.com/glfw/glfw/commit/05f6c13d119ea2662c97527d2421fb4cffd3dbfc

Currently, glfw3native.h of the internal GLFW is customized at
https://github.com/raysan5/raylib/commit/2feea87b616292b5bce4454a42c2d048f1cce7d8

This fix is compatible with the current customized glfw3native.h.

This fix enables us to update it to the latest and remove the
customization.