Changed the button logic
Before, you could hold mouse left button, drag it unto the button and change the sprite to "pressed" and, when released, the sfx would play. Now, you actually have to press the button in order to change the sprite to "pressed" and play the sfx.
It's a really minor change and maybe not welcomed, I only wanted better button logic on a program specifically made to show how to program button interaction.