Missing glsl1.2 shader for the Basic_Lighting example
If you run the /shaders/shaders_basic_lighting example on a platform with GLSL_VERSION set to 120, raylib isn't able to find the corresponding shader.
I merely copied the GLSL 1.0 version of the same shader from the other folder (and removed the precision mediump float line which was causing it not to compile) and pasted it in the glsl120 folder. It works well now.
Tested on RPi 0 2W running RPIOS Buster.