Allow excluding color constants from raylib.h
Add a macro, RL_NO_COLOR_CONSTANTS, to allow excluding the color macros
from raylib.h

Certain libraries conflict with raylib's color macros, and can't be
included in the same file

One library affected by this is reactphysics3d. the RED macro mangles
this particular line:
https://github.com/DanielChappuis/reactphysics3d/blob/17dd22e677ed861b0d4ece0c00a7e3cb503cc2f0/include/reactphysics3d/utils/DebugRenderer.h#L62
So, it cannot be used with raylib

Users may also choose to provide their own choice of color constants,
different to raylib's choice. This gives users the freedom to do so,
without modifying raylib.h

One practical example of this is raylib-namespace.h
https://github.com/raylib-extras/extras-cpp/blob/master/raylib_namespace.h
which prefers to define the color constants as constants instead of
macros