Add LoadModelEx and UploadModel functions
To enable custom model and mesh batching in projects that use the official raylib and instancing, it is essential to have control over whether or not meshes are automatically uploaded to the GPU. In my opinion, this is the cleanest solution for achieving that flexibility.

New functions added:
// Load model from files (mesh and material) without automatically uploading vertex data to the GPU Model LoadModelEx(const char *fileName, bool upload) // Upload model data to the GPU (RAM and/or VRAM)
void UploadModel(const Model *model, bool dynamic)

The existing LoadModel function now calls LoadModelEx with upload = true.

Additionally, fixed LoadOBJ to match the behavior of other load functions by not automatically uploading meshes.