Fix initial GetMouseDelta on GLFW
The initial mouse delta is often very large, depending on where the
cursor was before the window opened. This is because the first mouse
delta is the x/y coordinates your cursor entered the screen from. This
can (and often does) cause the free camera to make a sudden jump when
its moved for the first time.

The solution to this is to initialize the previousPosition and the
currentPosition to the same value on the first mouse input, essentially
ignoring the first mouse delta.