Update README.md
Copyedit English grammar.

In particular the common pattern of using a plural noun as a qualifier is ungrammatical in English but common among those learning English as a foreign language. There are rare exceptions like "materials system" and "releases page" so I didn't change those.

Please take this edit in good faith. I'm a native English speaker but also a learner of several languages in which I appreciate feedback on my persistent mistakes.